requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart
from_root
imbalanced-learn
Flask
gunicorn
openpyxl
apache-airflow
##-e .